MAJOR OVERHAUL
Code changes now allow for inclusion of absence data, and implements
creation of pseudoabsence (PAs) points as well. Additional option allows
for PAs to be picked only for areas beyond the climate envelope of all
available points for the species. In this case, number of PA points is
picked to reflect the density of P/A points within the climate envelope
of samples.
In terms of mechanics, now code runs from a single config file
(0_SDM_run_config) where all shared and step-specific configurations are
specified.